[bitset] Implement persistent bitset
The on-disk bitset consists of a series of bitmap blocks, indexed by
entries in a linked list, as they were considered outside the
transaction. The associated IndexEntry is similar to that of
dm-persistent-data, which stores the address of a bitmap block, and
tracks the number of enabled bits in each bitmap block.